fix(wre): validate trusted binding digest types
